refactor: move algorithm input code to io/
