+++ packages/web/src/state/algorithm/algorithm.state.ts
@@ -1,8 +1,8 @@
 import type { StrictOmit } from 'ts-essentials'
 import type { AuspiceJsonV2 } from 'auspice'
 
+import type { Virus, AnalysisResultWithMatch, AnalysisResult } from 'src/algorithms/types'
 import type { Sorting } from 'src/helpers/sortResults'
-import type { AlgorithmParams, AnalysisResultWithMatch, AnalysisResult } from 'src/algorithms/types'
 import type { QCFilters } from 'src/filtering/filterByQCIssues'
 import { getVirus } from 'src/algorithms/defaults/viruses'
 
@@ -48,6 +48,41 @@ export interface ResultsFilters extends QCFilters {
   aaFilter?: string
   cladesFilter?: string
   sorting?: Sorting
+}
+
+export enum AlgorithmInputType {
+  File = 'FileInput',
+  Url = 'Url',
+  String = 'String',
+}
+
+export interface AlgorithmInput {
+  type: AlgorithmInputType
+  name: string
+  description: string
+
+  getContent(): Promise<string>
+}
+
+export interface AlgorithmParams {
+  raw: {
+    seqData?: AlgorithmInput
+    auspiceData?: AlgorithmInput
+    rootSeq?: AlgorithmInput
+    qcRulesConfig?: AlgorithmInput
+    geneMap?: AlgorithmInput
+    pcrPrimers?: AlgorithmInput
+  }
+  errors: {
+    seqData: string[]
+    auspiceData: string[]
+    rootSeq: string[]
+    qcRulesConfig: string[]
+    geneMap: string[]
+    pcrPrimers: string[]
+  }
+  seqData?: string
+  virus: Virus
 }
 
 export interface AlgorithmState {
